requirements.txt
fastapi
uvicorn
requests
pydantic
pydantic-settings
jinja2
python-dotenv
